refactor(server): extract movement and unit constants to separate files
- Create MovementConstants.ts for movement system constants and terrain costs
- Create UnitConstants.ts for unit type definitions
- Remove duplicated constants from UnitManager.ts
- Update all imports across codebase to use new constants files
- Improve code organization and DRY principles
