+++ apps/server/tests/game/CityManager.test.ts
@@ -1,5 +1,5 @@
 import { CityManager, BUILDING_TYPES } from '../../src/game/CityManager';
-import { UNIT_TYPES } from '../../src/game/UnitManager';
+import { UNIT_TYPES } from '../../src/game/constants/UnitConstants';
 
 // Get the mock from setup
 // eslint-disable-next-line @typescript-eslint/no-require-imports, @typescript-eslint/no-var-requires
